move into molecule
